% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{lda_item_search}
\alias{lda_item_search}
\title{lda_item_search}
\usage{
lda_item_search(lda, query)
}
\arguments{
\item{lda}{model trained on 221386 title documents `all_lang`}

\item{query}{character vector with query words}
}
\description{
model-based search for publications
}
