% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{lda_topic_items}
\alias{lda_topic_items}
\title{lda_topic_items}
\usage{
lda_topic_items(lda, topic = 1)
}
\arguments{
\item{lda}{model trained on 221386 title documents `all_lang`}

\item{topic}{number of topic whose items should be returned}
}
\description{
topic model-based search for publications
}
